Added new functions, improved bException
bException() now no longer expects both $code and $e since always either one of the two is specified, so now it only expects at least 2 or at most 3 arguments. $code can either be a scalar exception code, or an exception object.

Added ensure_value() that will ensure that the specified value is one of the values specified in the $enum array. If so, $value will be returned. If not, $default will be returned

Added return_with_rights() will return $with_rights if the current user has the specified rights. If not, $without_rights will be returned