Improved cli_log()
cli_log() will now analyze the message that must be logged. If is an array it will log all entries, if it is an bException object, it will log all bException::getMessages(), if it is a normal PHP Exception, it will log Exception::getMessage(). It its a normal scalar value, it will just log that value