Improvbed file_temp(), file_tree_execute()
file_temp() no longer has to option to specify a file name. Instead there is a create option (default true) that will create the temporary file name, or if not set, it will only return the path that can be used for the temporary file

file_tree_execute() will now check if the path is specified and if it is absolute. If either conditions are not met, an exception will be thrown