Implemented library auto installer
The auto installer uses system library ensure_installed() and install library install(). ensure_installed() will check if the library is installed by a given set of rules. If any of the rules fail, a callback will be executed. This callback should contain a call to install() with the installation rules for said librar